requirements.txt
certifi==2023.7.22
numpy>=1.26.0
Pillow>=10.0.0
python-dotenv>=1.0.0
httpx==0.23.3
supabase==1.0.3
customtkinter>=5.2.0
opencv-python>=4.8.0
dlib==19.24.8
face-recognition==1.3.0
scikit-image>=0.21.0 